refactor: rename template and replace mod config values
